Computes metrics relative to the distance to the border of the extreme event (in surface, 2D)

This function calculate the median distance, mean distance and standard deviation of distance to escape form an extreme event 'as the crow files' (but avoiding NA pixels, which are suppose to represent a place that cannot be crossed) through time for a given GPS position or for all pixels.

BEE.calc.escape(
  true_event_output,
  start_date = NULL,
  end_date = NULL,
  pixel,
  only_days_EE = FALSE,
  group_by_event = FALSE
)

Arguments

true_event_output

: Is the ouput of function BEE.id.extreme_events().

start_date

: Defines the beginning of the timeframe during which you want to analyse distance to escape. If no dates are provided, computation will be done on all days_provided.

end_date

: Defines the endding of the timeframe during which you want to analyse distance to escape. If no dates are provided, computation will be done on all days_provided

pixel

: Is a vector with x,y coordinates or a df of x,y coordinates or "all" if you want to compute distance metrics for all pixels in the raster provided through true_event_output.

only_days_EE

: = 'TRUE' if you want to perform calculations using all the days. Use account = 'FALSE' when you want to calculate only for the days when your reference point is experiencing an extreme event. By default, account_all_days is set on FALSE

group_by_event

: Allows to specify if you want your result by extreme event or for the whole time periode specified. "group_by_event" = TRUE : compute one mean distance per event. "group_by_event" = FALSE : mean distance will be based on distances from all days in the time frame (belonging to an extreme event or not), including all days belonging to this event event if they are a '0' day, TRUE : compute one mean distance accounting only for days that belong to an EE.

Note

All distance metrics are limited by the size of the Spatraster your are providing.