BEE.calc.metrics_point compute metrics through time for specific locations.

BEE.calc.metrics_point compute metrics through time for specific locations.

BEE.calc.metrics_point(
  Events_corrected,
  Values,
  GPS,
  start_date = NULL,
  end_date = NULL,
  time_lapse_vector = NULL,
  baseline_qt = baseline_qt90,
  baseline_mean = baseline_mean,
  group_by_event = TRUE
)

Arguments

Events_corrected

Is the list of data tables produced by: bee_calc_true_event (the second element of the output). For each pixel, it contains a data.table with dates in the rows and a column indicating whether it is a day belonging to a heatwave (1) or not (0).

Values

is the spatraster with the values of the studded parameter, through time and space.

GPS

is a data frame containing the positions for which you want to compute metrics. It must contain columns labelled 'x' and 'y', which should contain longitudes and latitudes, respectively.

start_date

First day on which you want to start computing metrics.

end_date

Last day on which you want to start computing metrics.

time_lapse_vector

a vector of time laps on which to compute mean evolution rate and variance of the studdied parameter.

baseline_qt

Spatraster of the 90th percentile baseline (or the 10th percentile baseline)

baseline_mean

Spatraster of the 90mean value baseline.

group_by_event

Whether you want an output summarise by extreme event or not. If not, you just get daily metrics.

Value

A list of dataframe (one per GPS point), each dataframe contains informations on the date of the extrem events, mean, median, max and min values, peak day, onset-rate, off-set rate, mean anomaly, maximum category etc. Categories are defined in Hobday et al. 2018.

Examples

# To be added