Summarise the outputs of the 'merge' function.

To summarise one of the daily outputs of the following fonctions from BEE package : bee.calc.metrics_point() ; bee.calc.metrics_morpho() ; bee.calc.escape() ; bee.data.merge(). Metrics values can be summarised per extreme event or over a sliding time window.

BEE.data.summarise(
  data = data,
  variable = variable,
  summarise_by = "extreme_event"
)

Arguments

data

the dataset containing the column you want to process (e.g. the ouput of BEE.data.merge(), as a list of data frame).

variable

the name of the column for which you want to compute mean, median etc. over a time window.

summarise_by

takes the following options:

"extreme_event" for each metric, the following will be computed for each extreme event: mean, median, variance or sd, minimum and maximum, sum, and the dates of the maximum and minimum values.

a numeric value indicating the time step for which you want to compute the mean, median, etc.

#'@return a list of dataframes, with one dataframe per pixel, with the dates in the rows and the values summarised.

Details

In addition to the mean value, the function provides the median, standard deviation, minimum and maximum values, sum, date of minimum and maximum values (within the time window), and date of minimum value (also within the time window).

Examples

# TO BE ADDED